Adicionado páginação para consulta de fotos

+++ AltaMontanha/Models/Persistencia/Abstracao/INoticiaDAO.cs
@@ -7,6 +7,6 @@ namespace AltaMontanha.Models.Persistencia.Abstracao
 {
 	public interface INoticiaDAO : IDAO<Dominio.Noticia>
 	{
-		IList<Dominio.Noticia> Pesquisar(Dominio.Noticia objeto, int qtde);
+		IList<Dominio.Noticia> Pesquisar(Dominio.Noticia objeto, short qtde);
 	}
 }
